+++ src/model/Field.js
@@ -74,12 +74,12 @@ class Field {
                 delete data['field_type'];
             }
             if (data.hasOwnProperty('add_time')) {
-                obj['add_time'] = ApiClient.convertToType(data['add_time'], 'String');
+                obj['add_time'] = ApiClient.convertToType(data['add_time'], 'Date');
 
                 delete data['add_time'];
             }
             if (data.hasOwnProperty('update_time')) {
-                obj['update_time'] = ApiClient.convertToType(data['update_time'], 'String');
+                obj['update_time'] = ApiClient.convertToType(data['update_time'], 'Date');
 
                 delete data['update_time'];
             }
@@ -206,13 +206,13 @@ Field.prototype['field_type'] = undefined;
 
 /**
  * The creation time of the field
- * @member {String} add_time
+ * @member {Date} add_time
  */
 Field.prototype['add_time'] = undefined;
 
 /**
  * The update time of the field
- * @member {String} update_time
+ * @member {Date} update_time
  */
 Field.prototype['update_time'] = undefined;