+++ CHANGELOG.md
@@ -8,6 +8,10 @@ For public Changelog covering all changes done to Pipedrive’s API, webhooks an
 
 ## [Unreleased]
 
+## [22.6.0] - 2024-04-05
+### Added
+- Added documentation for new endpoints `/deals/{id}/changelog`, `/persons/{id}/changelog` and `/organizations/{id}/changelog`.
+
 ## [22.5.0] - 2024-02-02
 ### Added
 - Added documentation for new endpoint `/deals/{id}/participantsChangelog`.
@@ -593,7 +597,8 @@ structure
 * Fixed `GET /goal/:id/results` error handling in case when there are no existing stages connected to specified goal
 * Fixed typo in lead example response (`crrency` to `currency`)
 
-[Unreleased]: https://github.com/pipedrive/api-docs/compare/v22.5.0...HEAD
+[Unreleased]: https://github.com/pipedrive/api-docs/compare/v22.6.0...HEAD
+[22.6.0]: https://github.com/pipedrive/api-docs/compare/v22.5.0...v22.6.0
 [22.5.0]: https://github.com/pipedrive/api-docs/compare/v22.4.0...v22.5.0
 [22.4.0]: https://github.com/pipedrive/api-docs/compare/v22.3.0...v22.4.0
 [22.3.0]: https://github.com/pipedrive/api-docs/compare/v22.2.0...v22.3.0